Added Tags System with Tests and Filtering, and Invite members to the group

+++ client/test/helper.js
@@ -1,67 +1,82 @@
 import supabase from "../src/functions/setup";
 
-export const removeAllUsers = async () => {
-  // Fetch all users
-  const {
-    data: { users },
-    error: adminError,
-  } = await supabase.auth.admin.listUsers();
+export const removeTestUsers = async (userIds) => {
+  if (!userIds || userIds.length === 0) return;
 
-  if (adminError) {
-    console.error("Error fetching user list:", adminError.message);
-    return;
-  }
+  for (const userId of userIds) {
+    // Delete test user from auth and users table
 
-  for (const user of users) {
-    const { error: deleteError } = await supabase.auth.admin.deleteUser(
-      user.id
-    );
+    // check if the user is in the database
+    const { data: usersData } = await supabase
+      .from("users")
+      .select()
+      .eq("user_id", userId);
 
-    if (deleteError) {
-      console.error("Error deleting user:", deleteError.message);
+    if (usersData.length === 0) {
+      continue;
     }
-  }
 
-  const { data } = await supabase.from("users").select();
+    const { error: deleteError } = await supabase.auth.admin.deleteUser(userId);
+    if (deleteError) {
+      console.error("Error deleting test user:", deleteError.message);
+    }
 
-  for (const user of data) {
     const { error: truncateError } = await supabase
       .from("users")
       .delete()
-      .eq("user_id", user.user_id);
+      .eq("user_id", userId);
 
     if (truncateError) {
-      console.error("Error truncating 'users' table:", truncateError.message);
+      console.error(
+        "Error deleting test user from 'users' table:",
+        truncateError.message
+      );
     }
   }
 };
 
-export const removeAllGroups = async () => {
-  const { data: groups } = await supabase.from("groups").select();
+export const removeTestGroups = async (groupIds) => {
+  if (!groupIds || groupIds.length === 0) return;
 
-  for (const group of groups) {
+  for (const groupId of groupIds) {
+    // Delete test group
     const { error: truncateError } = await supabase
       .from("groups")
       .delete()
-      .eq("group_id", group.group_id);
+      .eq("group_id", groupId);
 
     if (truncateError) {
-      console.error("Error truncating 'groups' table:", truncateError.message);
+      console.error("Error deleting test group:", truncateError.message);
     }
   }
 };
 
-export const removeAllSongs = async () => {
-  const { data: songs } = await supabase.from("songs").select();
+export const removeTestSongs = async (groupId) => {
+  if (!groupId) return;
 
-  for (const song of songs) {
+  // Delete songs related to test group
+  const { error: truncateError } = await supabase
+    .from("songs")
+    .delete()
+    .eq("group_id", groupId);
+
+  if (truncateError) {
+    console.error("Error deleting test songs:", truncateError.message);
+  }
+};
+
+export const removeTestSongTags = async (songIds) => {
+  if (!songIds || songIds.length === 0) return;
+
+  // Delete tags related to test songs
+  for (const songId of songIds) {
     const { error: truncateError } = await supabase
-      .from("songs")
+      .from("song_tags")
       .delete()
-      .eq("song_id", song.song_id);
+      .eq("song_id", songId);
 
     if (truncateError) {
-      console.error("Error truncating 'songs' table:", truncateError.message);
+      console.error("Error deleting test song tags:", truncateError.message);
     }
   }
 };
